[common/sled_store] refactor: move sled based store into a standalone crate; Some types are moved because the rust orphan rule:Only traits defined in the current crate can be implemented for arbitrary types
