+++ common/metatypes/Cargo.toml
@@ -7,6 +7,8 @@ publish = false
 edition = "2021"
 
 [dependencies]
+common-sled-store = {path = "../sled-store"}
+
 serde = { version = "1.0", features = ["derive"] }
 async-raft = { git = "https://github.com/datafuse-extras/async-raft", tag = "v0.6.2-alpha.14" }
