+++ common/metatypes/src/lib.rs
@@ -14,7 +14,6 @@
 
 //! This crate defines data types used in meta data storage service.
 
-use std::cmp::Ordering;
 use std::collections::HashMap;
 use std::collections::HashSet;
 use std::fmt;
@@ -24,6 +23,9 @@ use std::fmt::Formatter;
 pub use cluster::Node;
 pub use cluster::Slot;
 pub use cmd::Cmd;
+pub use common_sled_store::KVMeta;
+pub use common_sled_store::KVValue;
+pub use common_sled_store::SeqValue;
 pub use errors::ConflictSeq;
 pub use log_entry::LogEntry;
 pub use match_seq::MatchSeq;
@@ -47,46 +49,6 @@ mod raft_types;
 
 #[cfg(test)]
 mod match_seq_test;
-
-/// Value with a corresponding sequence number
-pub type SeqValue<T = Vec<u8>> = (u64, T);
-
-/// The meta data of a record in meta-store kv
-#[derive(Serialize, Deserialize, Debug, Default, Clone, Eq, PartialEq)]
-pub struct KVMeta {
-    /// expiration time in second since 1970
-    pub expire_at: Option<u64>,
-}
-
-/// Value of StateMachine generic-kv
-#[derive(Serialize, Deserialize, Debug, Default, Clone, Eq, PartialEq)]
-pub struct KVValue<T = Vec<u8>> {
-    pub meta: Option<KVMeta>,
-    pub value: T,
-}
-
-/// Compare with a timestamp to check if it is expired.
-impl PartialEq<u64> for KVValue {
-    fn eq(&self, other: &u64) -> bool {
-        match self.meta {
-            None => false,
-            Some(ref m) => match m.expire_at {
-                None => false,
-                Some(ref exp) => exp == other,
-            },
-        }
-    }
-}
-
-/// Compare with a timestamp to check if it is expired.
-impl PartialOrd<u64> for KVValue {
-    fn partial_cmp(&self, other: &u64) -> Option<Ordering> {
-        match self.meta {
-            None => None,
-            Some(ref m) => m.expire_at.as_ref().map(|exp| exp.cmp(other)),
-        }
-    }
-}
 
 #[derive(Serialize, Deserialize, Debug, Default, Clone, PartialEq)]
 pub struct Database {
