+++ common/sled-store/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+name = "common-sled-store"
+version = "0.1.0"
+description = "distributed meta data service"
+authors = ["Databend Authors <opensource@datafuselabs.com>"]
+license = "Apache-2.0"
+publish = false
+edition = "2021"
+
+
+[dependencies]
+common-exception = {path = "../exception"}
+common-tracing = {path = "../tracing"}
+
+anyhow = "1.0.44"
+async-raft = { git = "https://github.com/datafuse-extras/async-raft", tag = "v0.6.2-alpha.14" }
+byteorder = "1.1.0"
+lazy_static = "1.4.0"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+sled = { git = "https://github.com/datafuse-extras/sled", tag = "v0.34.7-datafuse.1",default-features = false }
+tempfile = "3.2.0"
+
+
+[dev-dependencies]
+common-uniq-id = {path = "../uniq_id"}
+common-runtime = {path = "../runtime"}
+common-metatypes = {path = "../metatypes"}
+
+pretty_assertions = "0.7"
+
