+++ common/sled-store/src/sled_serde.rs
@@ -12,13 +12,22 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use std::mem::size_of_val;
 use std::ops::Bound;
 use std::ops::RangeBounds;
 
+use async_raft::raft::Entry;
+use async_raft::storage::HardState;
+use async_raft::AppData;
+use async_raft::LogId;
+use byteorder::BigEndian;
+use byteorder::ByteOrder;
 use common_exception::ErrorCode;
 use serde::de::DeserializeOwned;
 use serde::Serialize;
 use sled::IVec;
+
+use crate::SledValueToKey;
 
 /// Serialize/deserialize(ser/de) to/from sled values.
 pub trait SledSerde: Serialize + DeserializeOwned + Sized {
@@ -94,3 +103,59 @@ fn bound_ser<SD: SledOrderedSerde>(v: Bound<&SD>) -> Result<Bound<sled::IVec>, E
     };
     Ok(res)
 }
+
+impl<T> SledSerde for Entry<T> where T: AppData + Serialize + DeserializeOwned + Sized {}
+
+/// Extract log index from log entry
+impl<T> SledValueToKey<u64> for Entry<T>
+where T: AppData
+{
+    fn to_key(&self) -> u64 {
+        self.log_id.index
+    }
+}
+
+impl SledSerde for HardState {}
+
+/// NodeId, LogIndex and Term need to be serialized with order preserved, for listing items.
+impl SledOrderedSerde for u64 {
+    fn ser(&self) -> Result<IVec, ErrorCode> {
+        let size = size_of_val(self);
+        let mut buf = vec![0; size];
+
+        BigEndian::write_u64(&mut buf, *self);
+        Ok(buf.into())
+    }
+
+    /// (de)serialize a value from `sled::IVec`.
+    fn de<V: AsRef<[u8]>>(v: V) -> Result<Self, ErrorCode>
+    where Self: Sized {
+        let res = BigEndian::read_u64(v.as_ref());
+        Ok(res)
+    }
+}
+
+/// For LogId to be able to stored in sled::Tree as a key.
+impl SledOrderedSerde for String {
+    fn ser(&self) -> Result<IVec, ErrorCode> {
+        Ok(IVec::from(self.as_str()))
+    }
+
+    fn de<V: AsRef<[u8]>>(v: V) -> Result<Self, ErrorCode>
+    where Self: Sized {
+        Ok(String::from_utf8(v.as_ref().to_vec())?)
+    }
+}
+
+impl SledSerde for String {
+    fn ser(&self) -> Result<IVec, ErrorCode> {
+        Ok(IVec::from(self.as_str()))
+    }
+
+    fn de<V: AsRef<[u8]>>(v: V) -> Result<Self, ErrorCode>
+    where Self: Sized {
+        Ok(String::from_utf8(v.as_ref().to_vec())?)
+    }
+}
+
+impl SledSerde for LogId {}
