+++ kvlocal/Cargo.toml
@@ -17,6 +17,7 @@ common-metatypes = {path = "../common/metatypes"}
 common-runtime = {path = "../common/runtime"}
 common-tracing = {path = "../common/tracing"}
 common-store-api = {path = "../common/store-api"}
+common-sled-store = {path = "../common/sled-store"}
 
 metasrv = {path = "../metasrv"}
