+++ kvlocal/src/local_kv_store.rs
@@ -23,6 +23,7 @@ use common_metatypes::KVMeta;
 use common_metatypes::MatchSeq;
 use common_metatypes::Operation;
 use common_runtime::tokio::sync::Mutex;
+pub use common_sled_store::init_temp_sled_db;
 use common_store_api::kv_apis::kv_api::MGetKVActionResult;
 use common_store_api::util::STORE_RUNTIME;
 use common_store_api::GetKVActionResult;
@@ -33,7 +34,6 @@ use common_tracing::tracing;
 use metasrv::configs;
 use metasrv::raft::state_machine::AppliedState;
 use metasrv::raft::state_machine::StateMachine;
-pub use metasrv::sled_store::init_temp_sled_db;
 
 /// Local storage that provides the API defined by `KVApi`.
 /// It is just a wrapped `StateMachine`, which is the same one used by raft driven meta-store service.
