+++ metasrv/Cargo.toml
@@ -27,9 +27,10 @@ common-metatypes = {path = "../common/metatypes"}
 common-planners = {path = "../common/planners"}
 common-profling = { path = "../common/profiling" }
 common-runtime = {path = "../common/runtime"}
+common-sled-store = {path = "../common/sled-store"}
+common-store-api-sdk = {path = "../common/store-api-sdk"}
 common-tracing = {path = "../common/tracing"}
 common-uniq-id = {path = "../common/uniq_id"}
-common-store-api-sdk = {path = "../common/store-api-sdk"}
 
 # Github dependencies
