+++ metasrv/src/lib.rs
@@ -26,4 +26,3 @@ pub mod executor;
 pub mod meta_service;
 pub mod metrics;
 pub mod raft;
-pub mod sled_store;
