+++ metasrv/src/tests/service.rs
@@ -17,6 +17,7 @@ use std::sync::Arc;
 use anyhow::Result;
 use common_runtime::tokio;
 use common_runtime::tokio::sync::oneshot;
+use common_sled_store::get_sled_db;
 use common_tracing::tracing;
 // use common_tracing::tracing::Span;
 use tempfile::tempdir;
@@ -28,7 +29,6 @@ use crate::configs;
 use crate::meta_service::GetReq;
 use crate::meta_service::MetaNode;
 use crate::meta_service::MetaServiceClient;
-use crate::sled_store::get_sled_db;
 
 // Start one random service and get the session manager.
 #[tracing::instrument(level = "info")]
@@ -165,7 +165,7 @@ pub async fn assert_meta_connection(addr: &str) -> anyhow::Result<()> {
 macro_rules! init_meta_ut {
     () => {{
         let t = tempfile::tempdir().expect("create temp dir to sled db");
-        $crate::sled_store::init_temp_sled_db(t);
+        common_sled_store::init_temp_sled_db(t);
 
         // common_tracing::init_tracing(&format!("ut-{}", name), "./_logs")
         common_tracing::init_default_ut_tracing();
